src/main/java/br/com/cvc/teste/calculadores/CalculadorTaxaTipoCAcimaDe10Dias.java
package br.com.cvc.teste.calculadores;

import br.com.cvc.teste.utils.CalculadoraUtils;

/**
 *
 * @author johngomes
 */
public class CalculadorTaxaTipoCAcimaDe10Dias extends CalculadoraTaxa {
    
    private final double PERCENTUAL = 8.0;

    public CalculadorTaxaTipoCAcimaDe10Dias(CalculadoraTaxa proximo) {
        super(proximo);
    }

    @Override
    public double calculaTaxa(double valor, long numeroDeDias) {
        if (numeroDeDias > 10 && numeroDeDias <= 20) {
            return CalculadoraUtils.calculaValorSobPercentual(valor, PERCENTUAL);
        }
        return -1;
    }

}
